+++ src/core/errors.ts
@@ -2,6 +2,7 @@ export class ValidationError extends Error {
   public constructor(message: string) {
     super(message);
     this.name = "ValidationError";
+    Object.setPrototypeOf(this, new.target.prototype);
   }
 }
 
@@ -9,5 +10,6 @@ export class PersistenceError extends Error {
   public constructor(message: string) {
     super(message);
     this.name = "PersistenceError";
+    Object.setPrototypeOf(this, new.target.prototype);
   }
 }